Restructuring, Refactor: Tab Indent, HardTab for test.py and forms.py
