+++ img2xl_app/forms/UploadFileForm.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+from django import forms
+
+
+class UploadFileForm(forms.Form):
+	file = forms.FileField(
+		label='Select file',
+		help_text='Supported formats: jpg, jpeg, png, pdf (max 10MB)'
+	)
