+++ img2xl_app/forms/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+import RegisterForm
+import UploadFileForm
